+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.5.29 (20246)
 - Exposed ba method/class initing in public C++ layer.
 - The 'restart' and 'shutdown' commands in the server script now default to immediate=True
+- Wired up 'clean_exit_minutes', 'unclean_exit_minutes', and 'idle_exit_minutes' options in the server config
 
 ### 1.5.28 (20239)
 - Simplified ba.enum_by_value()